Unified Rm and rRm into Delsong.
Rm (remove a song) and rRm (remove range [start, end) of songs)
are now both represented in Delsong: based on the no. of arguments
either one song is removed or the range of songs.

Also added tilde to the list of characters replaced in printed file list
(this starts to look ridiculous, sorry).

Note that I have a new mail address, that is reflected in this commit.